Done Attendance

+++ internal/domain/models/ak_radius.go
@@ -1,7 +1,7 @@
 package models
 
 type Ak_Radius struct {
-	RadiusId     int     `gorm:"primaryKey" json:"radius_id"`
+	RadiusId     int64   `gorm:"primaryKey" json:"radius_id"`
 	NameLocation string  `gorm:"type:varchar(70)" json:"name_location"`
 	Latitude     float64 `gorm:"type:decimal(9,6)" json:"latitude"`
 	Longitude    float64 `gorm:"type:decimal(9,6)" json:"longitude"`
